Externalise option commands
