+++ app/lib/common/toolkit/errors/inline_error_button.dart
@@ -1,5 +1,5 @@
 import 'package:acter/common/toolkit/errors/error_dialog.dart';
-import 'package:acter/common/utils/utils.dart';
+import 'package:acter/common/extensions/options.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
