+++ app/lib/common/widgets/scrollable_list_tab_scroller.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:acter/common/utils/utils.dart';
+import 'package:acter/common/extensions/options.dart';
 import 'package:flutter/material.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 import 'package:scrolls_to_top/scrolls_to_top.dart';
