+++ app/lib/features/chat/widgets/convo_card.dart
@@ -14,6 +14,7 @@ import 'package:flutter_matrix_html/flutter_html.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logging/logging.dart';
 import 'package:skeletonizer/skeletonizer.dart';
+import 'package:acter/common/extensions/options.dart';
 
 final _log = Logger('a3::chat::convo_card');
